+++ src/misc/macros.cxx
@@ -3570,9 +3570,9 @@ static void pQSYFM(std::string &s, size_t &i, size_t endbracket)
 	do_qsy(false);
 }
 
-struct rfafmd { int rf; int af; std::string mdname;
-	rfafmd(int a, int b, std::string nm) { rf = a; af = b; mdname = nm;}
-	rfafmd(int a, int b) {rf = a; af = b; mdname = active_modem->get_mode_name();}
+struct rfafmd { unsigned long long rf; int af; std::string mdname;
+	rfafmd(unsigned long long a, int b, std::string nm) { rf = a; af = b; mdname = nm;}
+	rfafmd(unsigned long long a, int b) {rf = a; af = b; mdname = active_modem->get_mode_name();}
 	rfafmd(){rf = af = 0; mdname = active_modem->get_mode_name();}
 };
 static std::queue<rfafmd> fpairs;
@@ -3585,9 +3585,9 @@ static void pQSY(std::string &s, size_t &i, size_t endbracket)
 	}
 
 	std::string mdname = active_modem->get_mode_name();
-	int rf = 0;
+	unsigned long long rf = 0;
 	int af = 0;
-	float rfd = 0;
+	double rfd = 0;
 	std::string sGoFreq = s.substr(i+5, endbracket - i - 5);
 	// no frequency(s) specified
 	if (sGoFreq.length() == 0) {
@@ -3603,8 +3603,8 @@ static void pQSY(std::string &s, size_t &i, size_t endbracket)
 			if (pos == std::string::npos) triad = sGoFreq;
 			else  triad = sGoFreq.substr(0, pos);
 			sGoFreq.erase(0, triad.length()+1);
-			sscanf(triad.c_str(), "%f", &rfd);
-			if (rfd > 0) rf = (int)(1000*rfd);
+			sscanf(triad.c_str(), "%lf", &rfd);
+			if (rfd > 0) rf = (unsigned long long)(1000*rfd);
 			if ((pos = triad.find(":")) != std::string::npos) {
 				triad.erase(0,pos+1);
 				if (triad.length())
@@ -3648,9 +3648,9 @@ static void pQSY(std::string &s, size_t &i, size_t endbracket)
 
 static void doQSY(std::string s)
 {
-	int rf = 0;
+	unsigned long long rf = 0;
 	int audio = 0;
-	float rfd = 0;
+	double rfd = 0;
 	std::string sGoFreq;
 	sGoFreq = s.substr(6, s.length() - 7);
 	// no frequency(s) specified
@@ -3659,9 +3659,9 @@ static void doQSY(std::string s)
 		return;
 	}
 	// rf first value
-	sscanf(sGoFreq.c_str(), "%f", &rfd);
+	sscanf(sGoFreq.c_str(), "%lf", &rfd);
 	if (rfd > 0)
-		rf = (int)(1000*rfd);
+		rf = (unsigned long long)(1000*rfd);
 	size_t pos;
 	if ((pos = sGoFreq.find(":")) != std::string::npos) {
 		// af second value
@@ -3677,7 +3677,7 @@ static void doQSY(std::string s)
 	if (rf && rf != wf->rfcarrier())
 		qsy(rf, audio);
 	else
-		active_modem->set_freq(audio);
+		active_modem->set_freq((double)audio);
 	que_ok = true;
 }
